table: int and opt int column, column resize bugfix

+++ demo/demo-lib/src/jvmMain/kotlin/zakadabar/demo/backend/lib/BuiltinBackend.kt
@@ -6,9 +6,14 @@
 package zakadabar.demo.backend.lib
 
 import io.ktor.routing.*
+import org.jetbrains.exposed.sql.JoinType
+import org.jetbrains.exposed.sql.andWhere
 import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.transaction
 import zakadabar.demo.data.builtin.BuiltinDto
+import zakadabar.demo.data.builtin.ExampleQuery
+import zakadabar.demo.data.builtin.ExampleResult
+import zakadabar.stack.backend.authorize
 import zakadabar.stack.backend.data.record.RecordBackend
 import zakadabar.stack.data.DataConflictException
 import zakadabar.stack.util.Executor
@@ -23,15 +28,22 @@ object BuiltinBackend : RecordBackend<BuiltinDto>() {
 
     override fun onInstallRoutes(route: Route) {
         route.crud()
+        route.query(ExampleQuery::class, ::query)
     }
 
     override fun all(executor: Executor) = transaction {
+
+        authorize(true)
+
         BuiltinTable
             .selectAll()
             .map(BuiltinTable::toDto)
     }
 
     override fun create(executor: Executor, dto: BuiltinDto) = transaction {
+
+        authorize(true)
+
         if (dto.stringValue == "conflict") throw DataConflictException("stringValueConflict")
 
         BuiltinDao
@@ -40,16 +52,55 @@ object BuiltinBackend : RecordBackend<BuiltinDto>() {
     }
 
     override fun read(executor: Executor, recordId: Long) = transaction {
+
+        authorize(true)
+
         BuiltinDao[recordId].toDto()
     }
 
     override fun update(executor: Executor, dto: BuiltinDto) = transaction {
+
+        authorize(true)
+
         BuiltinDao[dto.id]
             .fromDto(dto)
             .toDto()
     }
 
     override fun delete(executor: Executor, recordId: Long) = transaction {
+
+        authorize(true)
+
         BuiltinDao[recordId].delete()
+    }
+
+    private fun query(executor: Executor, query: ExampleQuery) = transaction {
+
+        authorize(true)
+
+        val select = BuiltinTable
+            .slice(
+                BuiltinTable.id,
+                BuiltinTable.booleanValue,
+                BuiltinTable.enumSelectValue,
+                BuiltinTable.intValue,
+                BuiltinTable.stringValue
+            )
+            .selectAll()
+
+        query.booleanValue?.let { select.andWhere { BuiltinTable.booleanValue eq it } }
+        query.enumSelectValue?.let { select.andWhere { BuiltinTable.enumSelectValue eq it } }
+        query.intValue?.let { select.andWhere { BuiltinTable.intValue eq it } }
+        query.stringValue?.let { select.andWhere { BuiltinTable.stringValue eq it } }
+
+        select.map {
+            ExampleResult(
+                recordId = it[BuiltinTable.id].value,
+                booleanValue = it[BuiltinTable.booleanValue],
+                enumSelectValue = it[BuiltinTable.enumSelectValue],
+                intValue = it[BuiltinTable.intValue],
+                stringValue = it[BuiltinTable.stringValue]
+            )
+        }
     }
 }
